requirements.txt
huggingface-hub=0.24.6
pillow=10.4.0
torch=2.4.0
tesseract=0.1.3
tokenizers=0.19.1
torchvision=0.19.0
transformers=4.44.2
pytesseract=0.3.13